backend/pyproject.toml
[project]
name = "samba-wave-backend"
version = "0.1.2"
description = "Samba Wave Backend API"
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
    "aiohttp>=3.12.15",
    "anthropic>=0.52.0",
    "anyio>=4.10.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "boto3>=1.40.39",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.29,<0.30",
    "greenlet>=3.2.4",
    "gunicorn>=23.0.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.0",
    "python-multipart>=0.0.20",
    "python-ulid>=3.1.0",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    "fal-client>=0.5.0",
    "pillow>=11.0.0",
    "rembg[cpu]>=2.0.70",
    "httpx>=0.27.0",
    "curl_cffi>=0.7.0",
    "onnxruntime>=1.18.0",
    "openpyxl>=3.1.0",
    "slowapi>=0.1.9",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "security: marks tests as security tests (SQL injection, etc.)",
    "performance: marks tests as performance benchmarks",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["backend"]
omit = [
    "backend/__init__.py",
    "*/tests/*",
    "tests/*",
    "**/tests/*",
    "conftest.py",
    "*/conftest.py",
    "**/conftest.py",
    "backend/utils/pdf/*",
    "backend/utils/s3.py",
    "backend/api/v1/routers/*",
    "backend/main.py",
    "backend/domain/auth/*",
    "backend/domain/curai/*",
    "backend/domain/admin/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.ruff]
# scripts/ 디렉토리는 일회성 유틸/마이그레이션 스크립트 모음 — CI 린트 대상 제외
# (단 entrypoint.sh, verify_schema.py 등 배포 필수 파일은 런타임 이미지에 포함됨)
extend-exclude = ["scripts"]

[tool.ruff.lint]
# SQLAlchemy에서 == None, == True는 정상적인 ORM 패턴
# E402: 조건부 import 패턴 허용
# E701/E731/E741: 스타일 이슈 허용
ignore = ["E711", "E712", "E402", "E701", "E731", "E741", "F841"]

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["E402"]

[dependency-groups]
dev = [
    "alembic>=1.17.0",
    "black>=25.9.0",
    "isort>=6.0.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.3",
]
# deploy trigger
